refactor: address existing checkstyle violations in bugsnag-react-native android code
Checkstyle violations for our existing java code style rules have been fixed, notably: indent level,
import order, lack of braces, missing default branch in switch, top-level class extraction to
separate files
